Major kotlin port & recode, bug fixes (#956)
* Major kotlin port & recode, ScriptAPI compatibility, bug fixes

ClickGUI showcase: https://youtu.be/1_lgAtLQhBQ, https://youtu.be/Gbzj6ME5dAE

Ported to Kotlin:
* AccountsConfig, ButtonElement, ClickGUI, ClickGui, ClickGuiConfig, ClientUtils, Element, FileConfig, FileManager, FontDetails, Fonts, FriendsConfig, GuiClientFixes, HudConfig, ModuleElement, ModulesConfig, Panel, RenderUtils, Spammer, Style, ValuesConfig, XrayConfig and all ClickGUI styles: LiquidBounceStyle, NullStyle, SlowlyStyle

Recoded ClickGui - heavily changed all of ClickGui implementations, code and simplified everything with Kotlin.

General ClickGui improvements:
* fixed mouse interacting with all of overlapping panels, buttons, ... instead of with just the top one
* fixed panels being movable beyond the window border, now it even automatically adjusts when window size changes
* fixed ClickGui only showing fading animation when first opened
* fixed sliders having their slider over the limit when value was manually set above maximum
* added fade speed slider
* unified all right click interactions to play random.bow, not just Panel opening, closing

LiquidBounceStyle improvements:
* made scrollbars scale depending on how much there is to scroll
* added borders to module values

NullStyle improvements:
* made panel headers color to be inverted color from clickgui color, so that the text would remain readable with bright colors

Module utils:
* added moduleManager[moduleName] - works like moduleManager.getModule(moduleName)
* added module[valueName] - works like module.getValue(valueName)
- replaced all old getModule(), getValue() in code

Value utils:
* added FontValue.next(), FontValue.previous()
* added Value.range -> range from value.minimum to value.maximum
- used in ClickGui, HUD editor

Made EventManager, FileManager, ScriptManager and more constants class objects.
- other classes can import straight from them
-> replaced all eventManager.callEvent and such to just callEvent
-> LiquidBounce.fileManager.saveConfig(LiquidBounce.fileManager.valuesConfig) and such to just saveConfig(valuesConfig)
-> same with scriptManager, all LB fields, ...

SettingsUtils:
* fixed `.localconfig load` not being able to load TextValues with spaces

ScriptAPI compatibility:
* added @JvmStatic to make RotationUtils.setTargetRotation() and other utils work using ScriptAPI without having to access RotationUtils.INSTANCE
* didn't repeat the same mistake again with RenderUtils, ClientUtils, FileManager

* More improvements and final cleanup

Removed ALL unnecessary @JvmField and @JvmStatic
- it will break some scripts, but they were already broken with b77 RotationUtils port
TODO: utils could be provided with new api objects?

* Refactored `MovementUtils.isMoving`, `ClientUtils.displayChatMessage` and such to just `isMoving`, ... across all modules.

MovementUtils:
* Improved MovementUtils.speed: you can now do speed *= 1.08 and it will essentially call strafe(speed * 1.08) -> simplifies some Speed modules

ColorUtils:
* Implemented optional parameters into rainbow function.

BungeeCordSpoof:
* converted to object class

Sprint:
* fixed horizontal and vertical values isSupported override for newly introduced Legit mode

FriendsConfig:
* simplified serialization, deserialization
* removed old config support

New ClickGUI Panels:
* only move panels to windows border when mouse button is held
  -> prevents them from being moved to center when ClickGUI is opened in smaller window

Code improvement:
* Replaced PrintWriters to Kotlin's file.writeText.
* Replaced .equals to just == across all modules.
  -> simplified some if else blocks to when blocks
* Replaced Random().nextInt and such to just nextInt() from RandomUtils, Random.nextBoolean() -> nextBoolean(), Math.random() * ... -> nextFloat(min, max) ...
  + made parameters for RandomUtils functions optional
* Simplified some function blocks `{}` to just `= ...`
* Removed some unnecessary type annotations, null-safe types

MSTimer:
* hasTimePassed: Removed unnecessary casts to Long across all modules

ProphuntESP:
* fixed box, otherbox modes?
* impossible condition `(!mode.equals("Box", true) || !mode.equals("OtherBox", true))` -> && instead of ||

Updated Kotlin to 1.8.10

* Additional improvements

Improved GuiContributors:
* automatically selects first entry when opened
* navigate with arrows, tab, shift+tab, esc
* simplified with kotlin getter, ...

Improved AltManager (similar to GuiContributors)
* simplified navigation code
* added tab, shift+tab support

* Simplified floats, removed unnecessary .toFloat(), ...

* Additional stuff similar to previous changes... (imported static methods and fields from GlStateManager, RenderUtils, ...)

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/player/AntiAFK.kt
@@ -92,11 +92,11 @@ class AntiAFK : Module() {
                             delayTimer.reset()
                         }
                         4 -> {
-                            thePlayer.fixedSensitivityYaw += RandomUtils.nextFloat(-180.0F, 180.0F)
+                            thePlayer.fixedSensitivityYaw += RandomUtils.nextFloat(-180f, 180f)
                             delayTimer.reset()
                         }
                         5 -> {
-                            thePlayer.fixedSensitivityPitch += RandomUtils.nextFloat(-10.0F, 10.0F)
+                            thePlayer.fixedSensitivityPitch += RandomUtils.nextFloat(-10f, 10f)
                             delayTimer.reset()
                         }
                     }
@@ -108,13 +108,13 @@ class AntiAFK : Module() {
                 if (jumpValue.get() && thePlayer.onGround)
                     thePlayer.jump()
 
-                if (rotateValue.get() && delayTimer.hasTimePassed(rotationDelayValue.get().toLong())) {
+                if (rotateValue.get() && delayTimer.hasTimePassed(rotationDelayValue.get())) {
                     thePlayer.fixedSensitivityYaw += rotationAngleValue.get()
                     thePlayer.fixedSensitivityPitch += RandomUtils.nextFloat(0F, 1F) * 2 - 1
                     delayTimer.reset()
                 }
 
-                if (swingValue.get() && !thePlayer.isSwingInProgress && swingDelayTimer.hasTimePassed(swingDelayValue.get().toLong())) {
+                if (swingValue.get() && !thePlayer.isSwingInProgress && swingDelayTimer.hasTimePassed(swingDelayValue.get())) {
                     thePlayer.swingItem()
                     swingDelayTimer.reset()
                 }
